+++ app/lib/etl-toolkit/queues/populateIds.js
@@ -1,5 +1,6 @@
 const async = require('async');
 
+const processedPagesFilename = require('../../config').processedPagesFilename;
 const etlStore = require('../etlStore');
 const log = require('../../logger');
 const fsHelper = require('../../fsHelper');
@@ -13,12 +14,12 @@ function pageDone(pageNo) {
 
 function saveState() {
   etlStore.saveState();
-  fsHelper.saveJsonSync(processedPages, 'processedPages');
+  fsHelper.saveJsonSync(processedPages, processedPagesFilename);
 }
 
 function clearState() {
   processedPages = {};
-  fsHelper.saveJsonSync(processedPages, 'processedPages');
+  fsHelper.saveJsonSync(processedPages, processedPagesFilename);
 }
 
 function loadState() {
@@ -26,7 +27,7 @@ function loadState() {
   // if the ETL is cancelled between adding the IDs to the etlStore, and
   // setting the processedPages. Duplicate IDs in the list will be ignored in the next queue
   // so this will not cause problems
-  processedPages = fsHelper.loadJsonSync('processedPages') || {};
+  processedPages = fsHelper.loadJsonSync(processedPagesFilename) || {};
 }
 
 function handleError(err, pageNo) {